+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## v0.0.3
 ### Changed
- - Added sensor last_index
+ - Added sensor last_index compatible with energy dashboard
 ## v0.0.2
 ### Changed
  - Added sensor monthly_consumption